Add framework for quick restarts.
The Python interpreter is now in a separate module from
FRC_UserProgram.out (RobotPy.out).  FRC_UserProgram.out starts
the Python interpreter process and monitors it to see if it dies
and also monitors all module loading.

If it dies, the Python interpreter module and any loaded Python
modules are reloaded by FRC_UserProgram.out and the interpreter
is restarted.

A new builtin exception, SystemRestart, has been added to the
Python interpreter.  The interpreter process exits if this is
raised (any other exception will result in the interpreter
process hanging instead of exiting).